:truck: refactor: move httpHelper to a folder
