+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:55530/jest?","mongoDBName":"jest"}
+{"mongoUri":"mongodb://127.0.0.1:55577/jest?","mongoDBName":"jest"}
