+++ src/presentation/controllers/signup/signup.ts
@@ -1,5 +1,5 @@
 import { HttpResponse, HttpRequest, Controller, AddAccount, Validation } from './signup-protocols'
-import { badRequest, serverError, ok } from '../../helpers/http-helpers'
+import { badRequest, serverError, ok } from '../../helpers/http/http-helpers'
 
 export class SignUpController implements Controller {
   private readonly addAccount: AddAccount
